requirements.txt
flask
flask-cors
yt-dlp
openai-whisper
anthropic
python-dotenv
gunicorn
requests
